+++ internal/services/message/message.go
@@ -1,18 +1,24 @@
 package message
 
 import (
+	"github.com/CP-RektMart/pic-me-pls-backend/internal/chat"
 	"github.com/CP-RektMart/pic-me-pls-backend/internal/database"
+	"github.com/CP-RektMart/pic-me-pls-backend/internal/middlewares/authentication"
 	"github.com/go-playground/validator/v10"
 )
 
 type Handler struct {
-	store    *database.Store
-	validate *validator.Validate
+	store          *database.Store
+	chatSystem     *chat.Server
+	authentication authentication.AuthMiddleware
+	validate       *validator.Validate
 }
 
-func NewHandler(store *database.Store, validate *validator.Validate) *Handler {
+func NewHandler(store *database.Store, authentication authentication.AuthMiddleware, chatSystem *chat.Server, validate *validator.Validate) *Handler {
 	return &Handler{
-		store:    store,
-		validate: validate,
+		store:          store,
+		chatSystem:     chatSystem,
+		authentication: authentication,
+		validate:       validate,
 	}
 }
